sentenceAnagrams function improvement
sentenceAnagrams function re-implement to produce sentence anagrams in the forms of Word Group. For example one result of the function (with the sentence: "I love you") is [["olive"],["you"]] sentence anagram in the form of word group.  All sentence anagrams in the form of word group is following:
[["Io"],["Lev"],["you"]]
[["love"],["i"],["you"]]
[["olive"],["you"]]
Still improvements are necessary to obtain pure anagram sentences.