prefix implementation
prefix is implemented to get all words with given prefix. The method calls getWords function for subtree. Subtree means that the children of the last element in the prefix. It is obtained with traversing through the Trie with given prefix until reach the end. If it is not matched with the words in the Trie, then return Nothing.