update on insert
insert function is re-implemented to change "end flag" logic. "end flag" is set for the character placed in the before the "end flag". For example the "end flag" for the word "an" is set like following:
Trie {end = False, children = fromList [('a',Trie {end = False, children = fromList [('n',Trie {end = True, children = fromList []})]}

But in before implementation, "end flag" is set for the character placed in the after the "end flag". Like following:
Trie {end = False, children = fromList [('a',Trie {end = True, children = fromList [('n',Trie {end = True, children = fromList []})]}
So this leads to confusings for end of the word in some cases. This bug is fixed.